Move some methods extensions specs to its own folder
