+++ spec/extensions/params_helper_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative "../lib/params_helper"
-require "rspec_helper"
+require_relative "../rspec_helper"
+require_relative "../../lib/params_helper"
 
 RSpec.describe "ParamsBuilder" do
   it "builds params with colons" do
