+++ spec/extensions/presence_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require "rspec_helper"
-require_relative "../initialize/presence"
+require_relative "../rspec_helper"
+require_relative "../../initialize/presence"
 
 RSpec.describe "BaseExtensions" do
   describe :blank? do
